+++ serverside/player.js
@@ -33,6 +33,6 @@ Player.prototype.toArr = function(){
 }
 
 Player.prototype.toTypedArr = function(){
-  return new Uint8Array(this.toArr())
+  return new Uint16Array(this.toArr())
 }
 module.exports = Player
